structure clean up
