+++ framework/packages/abstract-sdk/src/lib.rs
@@ -29,7 +29,7 @@ pub use crate::apis::{
 pub use account_action::AccountAction;
 
 #[cfg(feature = "stargate")]
-pub use crate::apis::{distribution::*, stargate::feegrant::*};
+pub use crate::apis::{distribution::*, feegrant::*};
 
 pub mod features {
     //! # Feature traits
